+++ kernel/include/arch/x86/pit.h
@@ -6,7 +6,7 @@
 
 #define OSCILLATOR_FREQUENCY 1193182
 
-size_t PIT_init(
+uint32 PIT_init(
     uint16 a_divisor
 );
